Implement xen hypervisor console
Passing '--hvc' to bfexec will inform microv the guest is to use
the console=hvc0. The guest driver will do the proper initialization
and hypercalls to perform reads and writes to the console.

For now the guest only has the root user, and no password is needed
for successful login. Note that --hvc is mutually exclusive to both
--uart and --pt_uart.