Add XUE_DEBUG compile switch
By default, xue waits for the DbC to be configured before
continuing execution. This can cause the system hang if the debug cable
is not plugged in.

Add a XUE_DEBUG compile-time switch in bfdriver/src/common.c and
bfvmm/src/debug/unistd.cpp to allow the user to toggle xue usage
if desired.