xen: Stub BUG definition out when VMM is defined
The hvm/save.h public xen header includes an internal xen/lib.h
header that defines the BUG macro. This brings in stdargs and types
which causes redefinition errors in microv.

Check if VMM is #defined; if it is we are building bareflank so
don't include xen/lib.h and stub the definition for BUG.