Add domflags arg to create_domain vmcall
The create_domain domain op now accepts a flags argument in rbx. The
flag argument it the bitwise-or of the DOMF_* flags defined in
bfhypercall.h. These flags tell microv about how to treat the domain,
including things like exec mode and if it uses the xen console.

This allows for VMM resources associated with the flag bits to be
allocated in the constructor of the domain rather than after
construction with separate vmcalls.