Implement XSAVE support for host and guest vcpus
Each vcpu now has an 'xstate' object that manages the XSAVE
state components that are in use. The host vcpu's xstate is created
whenever a guest vcpu is created via the host's vcpu_op vmcall.

Every run_op, prior to the guest being run, the vmm saves the host
vcpu's xstate and loads the guest's. When the guest returns to the host
the vmm saves the guest state and loads the host's.

Note that since the vmm manually preserves each vcpu's SSE state, the
xstate save/load functions do *not* touch SSE state.