Add default host PCI handlers
When at least one PCI device is being passed-through to a guest VM, we
need to perform emulation of that device's PCI config space in both the
guest and host vcpus. This requires trapping the PCI IO ports. For each
device that is host owned (i.e., not being passed-through to a guest),
we add default portio handlers that passthrough the access to the
underlying device.