Implement host-side bits of PCI passthrough
Right now PCI passthrough is hardcoded to passthrough network devices to
domUs. In order to do this, each network device has PCI configuration
handlers installed at bootup. The handlers present a vendor:device of
0xBFBF:0xZZZZ, where ZZZZ starts at zero and increases with each device.
Drivers in the host wishing to route interrupts to the domU can use
these vendor:device IDs to bind to the devices.

Each capability is hidden from the host except for the MSI capability.
This is so the driver can request a vector from the kernel that can
later be routed to the domU.