Read VMCS rather than vcpu->cr4() in xstate init
Since the host VM owns most bits in cr4, writes to them do
not trap. This means that the software value vcpu->cr4() becomes
stale and does not reflect the actual value the guest is using.

We expects the host vcpu to have set cr4.osxsave prior to creating
the xstate area. Check the VMCS guest_cr4 rather than vcpu->cr4()
to ensure the right value is read.