Add lapic class and MSI mapping code
Each passthrough device installs PCI config space handlers in both the
guest and the host. The host handlers present a PCI vendor equal to
MICROV_PCI_VENDOR so that the host kernel will bind the microv kernel
driver. The driver's probe callback enables MSIs which causes the
kernel to program the device's MSI capability. The VMM records the MSI
address and data and adds it to the PCI device. The guest handlers
passthrough the vendor/device so that the guest kernel loads the actual
driver and also records the MSI information.

Once the guest enables MSIs on the device, a mapping is created in the
guest's parent host vcpu. This mapping may be used later by interrupt
handlers to map the host interrupt to the vector that should be queued
for injection into the guest.

In order to determine which host vcpu the mapping should be placed at,
the VMM must comprehend the destination ID field of the host MSI
address. The meaning of that field depends on the destination mode of
the local APIC's ICR.

- Add a local APIC class to determine the destination mode of the ICR on
each host vcpu.
- Add struct msi_desc for recording addr/data info programmed by a vcpu.
- Create a host -> guest MSI mapping when a guest enables MSIs.