hypervisor: Add push_external_interrupt
Interrupts may arrive at a different vcpu/pcpu than the one the guest VM
is running on. In this case, the vcpu that received the interrupt needs
to be able to queue the interrupt into the guest VM without having to
first load the VMCS.

Add push_external_interrupt to add the interrupt to the guest's queue
without accessing the VMCS. Note the disadvantage of using this function
is that if interrupt window exiting is disabled, the interrupt will sit
in the queue until window exiting is enabled, i.e. until someone
queue_external_interrupt()s the guest.