Add event_op hypercalls
Drivers in the host vcpus can use the event_op hypercalls to notify the
vmm that an interrupt has arrived for a passthrough device. The two
calls are:

    __enum_event_op__send_vector
    __enum_event_op__send_bdf

The first passes the host vector of the interrupt in rcx. The second
passes the bus/device/function of the device that generated the
interrupt in rcx.

The vmcall handlers use the rcx value to lookup the associated guest MSI
descriptor and queues the vector contained in the descriptor to the
guest's interrupt queue.