Create dom0 with the domain manager
Prior to this change, dom0 creation was bypassing the domain manager.
Instead a static instance was created the first time a vcpu was created.
Modify dom0 creation so it is the first domain created with the microv
domain manager. This enables users to acquire/release references to dom0
with get_domain/put_domain.