Change xen id generation
Each XENPVH guest has a set of ids that are different from the microv ids.
This change ensures that each xen vcpuid, apicid, and acpiid is 0 (since
each xen domain only has on vcpu - smp is not supported atm). If the
guest is a xenstore domain, the the domain id is 0, otherwise it is > 0