hypervisor: Add bfmanager::call_on_each
bfmanager::call_on_each calls the provided function on each t that it is
managing. The function must take a D* as the first argument, where D is
derived from T (T being the type that bfmanager was instantiated with).
In more precise terms, std::is_base_of_v<T, D> must be true in order for
the call to compile.