Add class xen_cpupool
Xen tools make several hypercalls related to the cpupool of a given
domain. This patch adds a class xen_cpupool to maintain the minimum info
necessary to implement these hypercalls. Each time a new xen_domain is
constructed, it is added to the default cpupool (the one with id = -1).
The tools read the cpupool information with xen_sysctl_cpupool_op or
xen_domctl_getdomaininfo. Tools may move a domain to another cpupool
with xen_sysctl_cpupool_op.