Add return_new_domain run_op return
When an xl-created domain is ready to run, the hypervisor returns
back to the root vcpu with the id of the new domain. The ID is used
in the vcpu-run loop of bfexec to create a new vcpu. This vcpu is
added as a child to the VM created from the bfexec command line.
After it is created, a new vcpu-run thread is started and the
new domain begins execution.