acpi: Add generalized hide_acpi_table
Generalize hide_dmar to take any struct acpi_table as an argument. This
function copies the original ACPI table to a new page with the header
zeroed-out and remaps the GPA of the table to the new page. Note this
remapping only applies to the root domain.

Use hide_acpi_table to hide the MCFG and DMAR tables from the root
domain when booting from EFI. Hiding the MCFG effectively disables PCI
Express and thus avoids trapping on the ECAM page to emulate config
space. Hiding the DMAR table prevents the root domain from accessing the
IOMMUs on the system (and as an unfortunate side-effect prevents them
from using x2apic).