uvctl: Implement event-based child management
This change adds a signalling mechanism from the uvc_vcpu::run() thread
to its parent domain. The mechanism enables a vcpu to notify its parent
domain of events it receives from the vcpu execution. The events
implemented so far manage Xen PVH domains created within an existing
vcpu that is running the xl toolstack.

The typical flow starts with the toolstack performing a hypercall that
operates on some domain. The hypercall handler in the VMM does whatever
actions it deems necessary in response to the operation, then returns to
its uvc_vcpu::run() thread in the root vcpu. The thread notifies its
parent domain of the event and then resumes execution.

Note that the event is handled asynchronously from the notification,
so event handler in userspace needs to have a general idea of what
the VMM has already done.