uvctl: replace initdom option with xsvm
To start a VM as the xenstore VM, pass --xsvm to uvctl. This tells the
guest itself and the hypervisor that it is allowed to perform privileged
hypercalls (mainly used to map in a domain's memory while it is building it).

This change also modifies the command line given to xsvm guests to
enable pciback to seize devices for later passthrough.