xen_memory: map and back pages on allocation
When EPT is shared with an IOMMU, any device access to an address
that isn't paged in will cause an IOMMU fault, not an EPT violation.
This change removes the on-demand paging of every domain so that
there is no need to worry about non-paged entries causing DMA faults.
Later we can distinguish between guests that require DMA remapping and
those that don't. The latter can be safely on-demand paged.