xen_domain: Bind iommu in ctor and assign_device
Add xen_memory::bind_iommu to bind an IOMMU to a domain's memory
manager. This is done as soon as we know that the domain has been
assigned pass-through devices. This ensures that any subsequent page
table entries incorporate information from the IOMMU such as snoop
control capabilities and whether the IOMMU's page walk is coherent.