xen_memory: Update snoop bit if IOMMU is bound late
The IOMMU is bound either at xen_memory construction or when
a PCI device that resides under the IOMMU is assigned to the domain
by the xl toolstack. In the latter case, if the IOMMU supports
snoop control, any existing EPT entries need to be updated with
the SNP bit (bit 11) set.