deps/linux: change to submodule rather than subtree
Various files within linux are duplicates when checked-out on a
case-insensitive file system (Windows). This change replaces the
linux subtree with a submodule. This allows for clone on Windows
to succeed by not including the linux submodule. The linux dependency
is only used for building guest images, which is not supported on
Windows, so preventing checkout of linux on Windows is fine.