lapic: Send EOI in guest interrupt handler
If an interrupt arrives at a CPU when a guest vcpu is running, the
interrupt causes an exit to the VMM. If the interrupt vector is assigned
to a guest, then the VMM writes the EOI to the local APIC on behalf of
the guest. This is required because the guest kernel's EOI is sent to a
virtual APIC managed by the VMM and never reaches actual hardware.

This change also simplifies the MSI mapping code. Instead of trying
to interpret what subset of CPUs an MSI targets, the VMM now creates
a root->guest MSI mapping on each root vcpu. Some of these mappings may
never be used, but this should be harmless.