xen: Begin platform PCI device
The Windows PV drivers bind to the platform PCI device. This
device provides a signal to the drivers that the VM in which they
are running was launched from QEMU and that they can proceed to
make any Xen hypercalls needed to initialize and operate.

The configuration space presents an "upstream" platform device (there
are other variants used by XenServer and other downstream projects).
The values of the configuration space and IO ports are derived from
linux (see include/xen/platform_pci.h) and
QEUMU (see hw/i386/xen/xen_platform.c).