domain: Remove references to domain->m_vcpu
Remove the pointer to microv::intel_x64::vcpu in the microv domain
class. This pointer really was only intended to work for a guest domain
with one vcpu. As we add support for multi-vcpu domain, a single pointer
is not going to suffice.

This change removes locations that were using the m_vcpu member and for
now only stores microv vcpuid's in the domain. Later we can use this to
get pointers to the domain's vcpus as needed.