xen/memory: Add support for winpv drivers shared info
Modify the XENMAPSPACE_shared_info implementation to support Windows
PV domain mapping in the shared info page. Note that for now, nothing
really happens, since the page is already mapped in the root domain's
EPT at the time of the call.

The FDO module of the PV drivers first create a 2MB "hole" in the
guest-physical address space with XENMEM_decrease_reservation. The
first 4K page within this hole is used for the shared info page.