xen/hvm: Emulate xs evtchn allocation for winpv drivers
Typically, a Windows PV driver domain runs as an HVM guest that
is created by xl. In the process of creating the guest, xl
EVTCHNOP_alloc_unbound's event channel ports for the domain's
xenstore and console. But since in our case the Windows PV domain
is a root domain, it is not created nor started by xl, and so the
VMM needs to perform the allocations on behalf of the toolstack.

This change allocates unbound event channel ports whenever the
Windows PV driver domain queries the ports with HVMOP_get_param.