winpv/xenbus: Handle lack of global line interrupt
A QEMU-created Xen platform PCI device advertises a legacy INTA
interrupt to the Windows VM. However, when run in a root domain,
the platform device does not expose an INTA interrupt because it
would require emulating an IOAPIC and GSIs in a consistent manner
with the ACPI tables. This situation causes FdoAllocateInterrupt
to fail for the LevelSensitive variation.

This change modifies the evtchn init code in xenbus to succeed after
FdoAllocateInterrupt fails for the level-triggered interrupt, provided
all the per-vcpu edge-triggered (Latched) interrupts are successfully
allocated. Each of the vectors associated with these interrupts are
registered with the VMM as a per-vcpu event channel upcall vector.