xen/gnttab: Mapspace from winpv root domain
Refactor xen_gnttab so that it is agnostic to the backing of a
shared frame. For guest domains, the frame is from the VMM. For
root domains, the gfn is mapped in from the add_to_physmap->gpfn
parameter. This change supports both of these cases, with the common
code operating on the virtual address derived from either backing.