xen/hvm: Alloc reserved pages for winpv domain
Typically, the toolstack allocates the xenstore and console
pages prior to booting an HVM guest in which the Windows PV drivers
reside. In a root domain the PV drivers run prior to xl, so the
VMM needs to create these on behalf of xl. It also needs to grant
the toolstack VM access to these pages in the Windows PV domain.

This change adds these pages whenever a root domain with Windows
PV enabled is constructed and grants the toolstack VM access to
the pages whenever the first shared grant page is added with
XENMEM_add_to_physmap