xen/hvm: Add store/console pages into root domain
Whenever xs_introduce_domain is called by init-winpv in dom0, the
xenstore daemon maps the store grant reference in so that communication
can begin. This change adds the store (and console, for later use) page
to the domain with add_vmm_backed_page so that when the map_grant_ref
code later goes to look up the page it will be found.