xen/evtchn: Use shared_info to upcall for winpv
The Windows PV xenbus driver checks for
vcpu_info[vcpuid].evtchn_upcall_pending != 0 to kickoff the poll of the
FIFO queues. The currently supported version of Linux ignores this
whenever FIFO is being used.

Set evtchn_upcall_pending to 1 whenever an upcall is made to a guest vcpu.
The guest clears this before processing pending upcalls.