xen: Add evtchn and gnttab changes for starting Windows PV
The Windows xenbus driver binds a vcpu to VIRQ_DEBUG for debugging
purposes. To support this, check that the given vcpuid to bind is valid
and then set the channel's vcpuid to the given value.

Also during start, xenbus pings the dom0-side of xenstore and will
hang/timeout if it doesn't get a reply. Because of this, the dom0 VM
must be started and /usr/lib/xen/bin/init-winpv must be run within the
dom0 VM prior to starting xenbus. In order for init-winpv to succeed, an
exception is made (for now) in xen_gnttab_map_grant_ref when mapping the
xenstore page into the dom0 VM. At that point in execution, the xenbus driver
hasn't started, so the VMM hasn't mapped in xenbus's grant page, so the
xenstore ref is out-of-bounds. In this scenario though, we have all the info
we would typically read from the grant entry (i.e. if we have read/write
access and the gpfn, the latter being in the Windows domain's HVM parameters),
so we just use it without accessing the grant entry. Note though that
since we don't have the entry, we can't update GTF_reading|GTF_writing,
so we would need to set these flags whenever the xenbus driver adds the
grant page with XENMEM_add_to_physmap.