xen/gnttab: Add special case unmap for Windows PV
This change provides an exception in xen_gnttab_unmap_grant_ref for the
Windows PV domain that mirrors the xen_gnttab_map_grant_ref.  The
exception is made (for now) because it is possible the Windows PV domain
hasn't added the shared grant page yet. In that case for the xenstore
reference, we just go straight to unmap from the local domain because we
have all the info we need without the shared entry.