deps/xen: Add root_domain build info variable
Add a root_domain variable in the libxl_domain_build_info type
to distinguish a root domain from non-root domains. A root domain
is the domain the hardware is booted into (similar to dom0, but
deprivileged and treated as a domU in most cases from a toolstack
perspective). Right now, a root_domain is any domain with UUID
equal to 01000000-0000-0000-0000-000000000000.