vmm: Add support for "creating" root domain with xl
This change enables you to run xl create /etc/xen/<rootcfg>
for the root domain. The "root domain" is identified with a UUID
of uuid = "01000000-0000-0000-0000-000000000000" in <rootcfg>.
This tells xl that a root domain is being created, which will then
populate the xenstore entries for the domain so that later commands
like xl network-attach will work as expected.