winpv/xenbus: Drop store access to <= APC_LEVEL
Each store access submits a request with StoreSubmitRequest.  This
function waits for a response from the remote domain before continuing.
Priory to this change, this waiting happened at DISPATCH_LEVEL, which
caused deadlock whenever the submitting thread executed on the same CPU
as the remote domain. This is not a problem when Xen is used as the VMM
because Xen schedules VMs itself. However it is a problem when MicroV is
the VMM because the host OS (i.e. Windows) is in charge of scheduling.

The change modifies the store code to StoreSubmitRequest at <= APC_LEVEL
and yields the processor to Windows with KeDelayExecutionThread while
waiting for the response. Unfortunately, this change breaks every call
site that accesses the store at DISPATCH_LEVEL, so those will be
addressed in later commits.