vmm: Add support for root-vcpu event injection
This change adds functions to the xen_vcpu class for delivering
interrupts to the guest. The functions check if the vcpu is a root
vcpu, and if so, it sends an IPI through the vcpu's local APIC rather
than using the VMCS injection mechanism. This ensures that the EOI
from the OS running in the root domain corresponds to an actual
bit in the local APIC ISR. If the vcpu is a guest, then the standard
VMCS queueing/injection mechanism is used.