xen/gnttab: Add root domain exception in xen_gnttab_copy
If a guest does xen_gnttab_copy with a root domain as either the
source or destination, it is likely that the corresponding page hasn't
been associated with a class xen_page yet. In this case, we create
a xen_page in the root domain that is backed by the page frame given
in the copy argument.