winpv/xenbus: Remove ProcessorID from SYSTEM_PROCESSOR
The ProcessorID field from the local APIC entries in the MADT
were originally used as the vcpuid. A prior change removed this
in favor of identity mapping the vcpuid to the 0-based, linear index
provided by Windows. This ensures that the set of vcpuids starts at
0 (the ProcessorID field from the MADT does not necessarily start at 0).

Because of this change, the ProcessorID field is no longer used and so
the MADT is no longer needed. This change removes the call to
AcpiGetTable that maps in the MADT and thus prevents a bug in
AuxKlibGetSystemFirmwareTable returning STATUS_NOT_IMPLEMENTED randomly
during boot. Googling the above issue doesn't have many hits, but one
site did describe it being possible due to timing between the AuxKlib
the ACPI subsystem initialization. Windows doesn't provide a means
to synchronize the two, so we just side-step the problem entirely
by not using the ACPI tables.