uvctl: Ensure HVC Tx buffer is mapped
If the buffer passed to __domain_op__hvc_tx_get is not mapped into
uvctl's address space, the VMM will not be able to translate the virtual
address of the buffer to the corresponding guest-physical address.
This change attempts to lower the probability that the buffer is not
mapped in by reducing its size to one 4K page via HVC_TX_SIZE. It also
writes to the first byte just prior to performing the hypercall to try
to convince the OS that it needs to be mapped.

Obviously this still doesn't prevent the OS from paging it out, but the
goal is that the single page will be hot enough to evade the OS's page
replacement algorithm. In order to provide a guarantee, we would need
to use OS-specific facilities to lock the buffer into RAM.