vmm: Handle xenstore readiness vmcalls from xsvm
Add support for signaling that xenstore is ready from the xenstore VM
(i.e. xsvm or "dom0") itself. The Windows PV driver still queries until
it is signaled ready. Note that this change repurposes init-winpv to
signal that xenstore is ready. This can be added to a start up service
after the winpv domain has been created.