xen/evtchn: Rework event signalling
Convert the hand-rolled upcall algorithm to the one officially defined
in https://xenbits.xenproject.org/people/dvrabel/event-channels-F.pdf.

Since the Windows root domain is multi-core, it requires SMP-safe event
channel signalling, so the upcall code is revamped to use locking and
atomic accesses where appropriate. The changes ensure that
VMM-concurrent accesses to the FIFO queues and guest-concurrent accesses
to event words are synchronized.

Also, ensure that an event is raised on the local port allocated in
EVTCHNOP_bind_interdomain as is done in Xen's implementation.