vmm: Add root domain EPT handler
Add a catch-all EPT handler for the root domain. Right now the root
domain is identity-mapped into physical RAM so the handler
identity-maps the page RWE. Note that we dont implement TLB shootdowns
for now because of the added complexity and the fact that each core's
TLB entry for a given address will converge to the same value over time.