xen: Add raw page for root domain gnttab
Instead of adding a root backed page into the root domain
during xen_gnttab_map_grant_ref, add a "raw" page into the local
domain that is doing the map. A "raw" page is like a root-backed
page in that it is backed by a root host frame, but it is not from
the root pool (i.e. not allocated from the builder driver).

Now we ensure that xen_gnttab_map_grant_ref uses a raw page
for the root-domain memory so that when the mapping is unmapped,
we can release the page (this wasn't happening with the root-backed
pages).