vmm: Add missing invept calls
Ensure that each domain's EPT TLB entries are invalidated whenever its
mapping changes by calling invept(). Note that for hypercalls that
operate on a domain that is not the caller do an invept on the core the
*caller* is running on. Right now, this core is the same as the core of
the domain whose EPT is being invalidated so there is no need for an
IPI/TLB shootdown. Later when this assumption no longer holds (i.e.
guests run on different CPUs) we will need to perform the invept
on the proper physical core(s).