deploy: Add scheduled tasks and power config
Register 'StartVms' and 'StartNet' task on each install
and unregister the same on each uninstall. The former task boots
the service VMs and the latter configures the Xen PV network device
once it is detected.

Also modify the target's power configuration to disable sleep on
install and to re-enable sleep on uninstall.