vpnvm: Add DHCP server for Windows domain
Trying to assign a static IP to the Xen interface at every boot
is error prone. Instead, use a DHCP server that serves IPs on
the 192.168.5.0/24 subnet to the Windows domain. The DHCP server
is started with the dhcpd.service file which waits until the
vif interface is up (i.e. until the demo.service is done).