xen/evtchn: Add EVTCHNOP_reset and rework EVTCHNOP_close
Implement EVTCHNOP_reset hypercall. Note that so far this is called only
from the Windows PV xenbus driver; Linux guest VMs do not call this. The
intent of the implementation is for shutdown and not really reset. A
proper reset actually takes the event channel ABI back to two-level -
the current implementation does not do this, it just closes each channel
and returns. In order to support suspend/resume we will need to
implement this with the two-level fallback like Xen.

Rework event channel close so that it is locked wrt remote EVTCHNOP_sends.
Also allow for the upcall vector to be set to 0 by the HVMOP hypercalls.
The Windows xenbus driver does this when the event channels are being
torn down.