uvctl: Add Windows xenbus IOCTL interface
Define an IOCTL interface for the Windows xenbus driver.
The interface is defined with the GUID_DEVINTERFACE_XENBUS in
include/microv/xenbusinterface.h. This is used by uvctl to open
a file for performing IOCTLs to/from xenbus (in addition to the
builder driver's IOCTL interface).

Note that the driver-side part of the IOCTL interface will be
added in a later commit.