winpv/xenbus: Add xenstore process protection
This patch implements the kernel portion of the xenbus IOCTL interface.
It also uses the interface (really the corresponding IRP_MJ_CREATE sent
in response to userspace opening an fd on the interface) to implement
termination protection of the process that is providing the xenstore
backend VM.

Whenever a process opens a file descriptor to the xenbus interface, the
driver registers callbacks on the process's handle operation functions
with ObRegisterCallbacks. The kernel invokes these callbacks whenever
any subsequent handle is opened on the process. The callbacks remove the
handle's access rights to terminate the process, which effectively makes
the process unkillable.

As of now, the uvctl process opens a handle to the xenbus interface in
addition to the builder interface, which makes uvctl unkillable after
successfully opening the xenbus interface fd. This ensures that the
xenstore backend provided by VMs running from uvctl always exist for
xenbus to talk to. This protection is necessary because if the xenstore
backend is not there, xenbus threads doing xenstore IO will sleep
forever.