winpv/xenbus: Relax cache cursor ASSERT
Remove the ASSERT in CacheDestroySlab that checks if the Cache->Cursor
is being destroyed, then it is the only one in the list. Specifically,
remove the check on the Blink field of the slab entry. This ASSERT is
hit on occasion whenever network IO is taking place, but removing it
seems to be harmless.

The CacheAudit function states that the cursor should always point to
the first slab that is not fully empty. This check is performed in place
of the ASSERT with appropriate error messages indicating the check has
failed.