winpv/xenbus: Add evtchn fixes for FDO stop/removal
Whenever the xenbus FDO is stopped or removed, it releases
the EVTCHN interface, which in turn EvtchnFlush()es each event
channel and does an EvtchnReset.

Remove the ASSERT in EvtchnFlush that checks that
Processor->UpcallEnabled if the line interrupt (i.e. Context->Interrupt)
is NULL. This ASSERT is not necessary and just isn't true because each
UpcallEnabled is cleared in EvtchnInterruptDisable prior to calling
EvtchnFlush.

Also, remove the call to EvtchnReset. That function assumes the xenbus
driver is running in an HVM guest and does unnecessary calls for
re-binding the store and console event channels. In our case the
xenbus driver is running in a root domain which means that the
entire system is being shutdown so the implementation should assume
there isn't another domain to bind events to.