vpnvm: Signal when VPN is running
Add the xen tools to the vpnvm image so that the vpnvm may
xenstore-write "1" to /local/domain/2/data to signal the VPN is
up and running. The xsvm spins on /local/domain/2/data until it
reads 1. Then it proceeds to network-attach the Windows domain
to the netback in the vpnvm. This helps ensure that the network
is running before the xenvif in the Windows domain is started.