scripts: add make-installer.ps1
Add make-installer.ps1 which calls the Inno Setup compiler to create an
installer binary. It takes two optional parameters: IssPath and Name.
The first is the path to the Inno Setup script to compile. By default
this is deploy/windows/app.iss. The Name parameter is the desired name
of the installer binary, which is "install-beam" by default.

Also remove scripts/init-net.ps1 as it is no longer used.