Remove deps/linux submodule
Since we don't have any changes to upstream linux for our guest VMs
we don't need to maintain a submodule for linux (nor a subtree
for that matter). This patch moves the linux defconfigs of each VM to
config/linux/ and modifies the buildroot defconfigs to use the
upstream tarball for linux 5.4.13. While we're here, also delete
venom_basepvh_defconfig and venom_vm_config.in since they are unused.

Note if you really need to modify the kernel of a VM, you can still
point buildroot to a local repo by defining the LINUX_OVERRIDE_SRCDIR
variable in local.mk.in.