efi: add EFI_BOOT_NEXT config option
Add a cmake config variable "EFI_BOOT_NEXT" that contains the path to
the EFI binary that bareflank.efi should run after the VMM has started.
The path uses Windows-style '\' characters and is formatted as a
wide-character C-string and thus must be prefixed with "L". It is
interpreted as relative to the root of the EFI system partition (ESP).
So if your ESP is at /boot/efi and you want bareflank.efi to run
/boot/efi/EFI/grub/grubx64.efi, then following in your config.cmake
will give the proper value to EFI_BOOT_NEXT:

set(EFI_BOOT_NEXT "L\"\\\\EFI\\\\grub\\\\grubx64.efi\"")

Then be sure to configure cmake before running make and you'll be good
to go.