deploy/windows: add uefi shell option
This patch adds the pre-compiled shell from tianocore so
that it can be included in the installer if desired. Users
can opt to boot into the shell first rather than bareflank.efi
by suppling the -useshell switch to make-installer.ps1.
This enables command line options to be passed to bareflank.efi
which can be very useful for testing and development.