xen/evtchn: move event_control to xen_vcpu
Instead of keeping an array of struct event_controls in xen_evtchn,
give each xen_vcpu a struct event_control, and then look up
the vcpu in order to access event_control when needed when raising
events. This ensures that the event_control is tied to the lifetime
of the xen_vcpu instead of keeping track of them separate from the
domain's xen_evtchn.