vmm: use queue_external_interrupt when possible
This patch ensures that queue_external_interrupt is used when it can be
safely (as opposed to push_external_interrupt). Using the queued version
ensures that interrupt window exiting is enabled immediately as opposed
to waiting for another event to do it. This is only possible if the
target vmcs is active on the same physical CPU as the one where
queue_external_interrupt is called from.  Using the queued version is
preferred when latency matters, as this will keep the time between when
an interrupt arrives into the VMM and when it is injected into the guest
at a minimum.