drivers/winpv: fix build warning and xennet shutdown bug
Check for a NULL ListEntry in EvtchnDebugCallback to silence
the 'Dereferencing NULL pointer' warning.

Don't call AdapterDisable from xennet's MiniportShutdownEx. This
ensures that no xenstore access is made. Xenstore access is not
allowed from shutdown context because the process that is hosting
the xenstore backend can (and is) killed by the kernel in parallel to
MiniportShutdownEx. Any subsequent xenstore access will hang, causing
the shutdown itself to hang.