Fix: report TSC missing info from userspace
Fixes issue Bareflank/boxy#48

This removes the need to debug over serial this error.
We fail silently in the VMM and report the error in bfexec from
userspace. We can do this because the tsc freq feature is only
needed after the launch of a VM and we have the opportunity
to fail before from userspace.

A more robust way to do this might be to add an `environ_check`
executable which would be called from the `driver_load.sh` script
and in which this check is done along with others.
E.g. the `VMX extensions not supported` error has the same problem
of needing a serial to be debugged.