deploy/windows: add handler for vif connect events
This patch adds two files used for handling vif connect events.
vifctl.ps1 is run on each install/uninstall. On install it registers
two ScheduledTasks: VifConnect and NetctlWifi.

VifConnect is triggered whenever a new network connection comes online.
The trigger causes vifconnect.ps1 to be called, which then inspects the
state of the network interfaces and sets their properties. The most
important interface it looks for is the management interface (all
interfaces are assumed to have a fixed MAC, see vifconnect.ps1 for their
definitions). Whenever the management interface is connected, it disables
IPv4, enables file sharing, and sets its IPv6 address. It also sets the
name of the network so it is easy to tell networks apart from Windows`
GUIs. It then Starts the NetctlWifi task if it is not already running.
NetctlWifi simply runs the netctl-wifi.exe (although as committed, it
isnt being run because of issues with netctl-wifi.exe. Once those are
fixed we should be able to run from here, although it may be simpler
to just have the user click on it manually).

On uninstall, vifctl.ps1 unregisters the VifConnect and NetctlWifi
ScheduledTasks.