deploy/windows: add auth to startvms.ps1 and optional console
startvms.ps1 now accepts a $CredFile argument, which if given is
assumed to point to a file containing the user's username and password
(stored as a standard encrypted string) in the form

  user:<user>
  pass:<pass>

If $CredFile isn't given, the user is prompted for their credentials,
which are stored in $PSScriptRoot\cred.txt. The credentials are then
parsed out of the $CredFile and passed to systemd as environment
variables that are used to mount the samba share. It also reads
the value of the UVCTL_PCIBACK_HIDE environment variable (set by
setenv.ps1 during installation) and passes it to xen-pciback.hide=
on the xsvm's kernel command line.

If $Console is given, then uvctl boots the xsvm to an interactive
console which can be used for debugging. Otherwise it runs it
in the background.