acpi: tables mappings.
Keep track of the ACPI tables in the hypervisor, but do not map them
unless necessary. The granularity of contiguous regions where ACPI
tables can be found is reduced to make it easier to replace or hide some
tables in the guest.

Call sites where tables were accessed are now responsible for setting up
a valid mapping to said table in the hypervisor. It makes it easier to
have contiguous va to one table without having to keep track of what is
mapped where.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>