vmm: add g_enable_xue switch
This patch adds a command-line option to bareflank.efi called
"--enable-xue" which will tell the VMM to use the xue debugger, provided
that the VMM was built with USE_XUE set to ON in the config.cmake file.

When --enable-xue is provided, it sets the g_enable_xue variable to
1 defined in deps/hypervisor/bfdriver/src/common.c, which eventually
results in a call into the hypervisor with BF_REQUEST_INIT_XUE, which
sets the g_enable_xue in the VMM's address space to "true". All accesses
to xue in the VMM are guarded by g_enable_xue.

Note that the windows and linux loaders set g_enable_xue to 1
unconditionally, provided that USE_XUE is defined for the driver build.
This is because there is no command line, and this patch doesn't include
any changes to bfm, which would be the natural place to put this if we
think it is needed.